+++ src/apps/sequencer/model/NoteSequence.cpp
@@ -28,6 +28,8 @@ Types::LayerRange NoteSequence::layerRange(Layer layer) {
     CASE(Condition)
     case Layer::AccumulatorTrigger:
         return { 0, 1 };
+    CASE(PulseCount)
+    CASE(GateMode)
     case Layer::Last:
         break;
     }
@@ -70,6 +72,10 @@ int NoteSequence::layerDefaultValue(Layer layer)
         return int(step.condition());
     case Layer::AccumulatorTrigger:
         return step.isAccumulatorTrigger() ? 1 : 0;
+    case Layer::PulseCount:
+        return step.pulseCount();
+    case Layer::GateMode:
+        return step.gateMode();
     case Layer::Last:
         break;
     }
@@ -107,6 +113,10 @@ int NoteSequence::Step::layerValue(Layer layer) const {
         return int(condition());
     case Layer::AccumulatorTrigger:
         return isAccumulatorTrigger() ? 1 : 0;
+    case Layer::PulseCount:
+        return pulseCount();
+    case Layer::GateMode:
+        return gateMode();
     case Layer::Last:
         break;
     }
@@ -157,6 +167,12 @@ void NoteSequence::Step::setLayerValue(Layer layer, int value) {
         break;
     case Layer::AccumulatorTrigger:
         setAccumulatorTrigger(value != 0);
+        break;
+    case Layer::PulseCount:
+        setPulseCount(value);
+        break;
+    case Layer::GateMode:
+        setGateMode(value);
         break;
     case Layer::Last:
         break;